backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
openai>=1.50.0
python-dotenv==1.0.1
bleach==6.3.0
pydantic-settings==2.7.0
asyncpg==0.30.0